% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Colar dois vetores}
\usage{
colar(a, b)
}
\arguments{
\item{a}{primeiro vetor}

\item{b}{segundo vetor}
}
\value{
Um novo vetor colado, com um espa??o entre eles
}
\description{
Cola dois vetores e mant??m um espa??o entre eles
}
